+++ src/components/Skills.jsx
@@ -1,54 +1,11 @@
 import { CircularProgressbar } from "react-circular-progressbar";
 import "react-circular-progressbar/dist/styles.css";
-import { useEffect, useReducer } from "react";
-import axios from "axios";
-import {
-  skillReducer,
-  initialState,
-  actionTypes,
-} from "../reducers/skillReducer";
 import { requestStates } from "../constants";
+import { useSkills } from "../customHooks/useSkills";
 
 export const Skills = () => {
-  const [state, dispatch] = useReducer(skillReducer, initialState);
-
-  useEffect(() => {
-    dispatch({ type: actionTypes.fetch });
-    axios
-      .get("https://api.github.com/users/shun-naganuma/repos")
-      .then((response) => {
-        const languageList = response.data.map((res) => res.language);
-        const countedLanguageList = generateLanguageCountObj(languageList);
-        dispatch({
-          type: actionTypes.success,
-          payload: { languageList: countedLanguageList },
-        });
-      })
-      .catch(() => {
-        dispatch({ type: actionTypes.error });
-      });
-  }, []);
-
-  const generateLanguageCountObj = (allLanguageList) => {
-    const notNullLanguageList = allLanguageList.filter(
-      (language) => language != null
-    );
-    const uniqueLanguageList = [...new Set(notNullLanguageList)];
-
-    return uniqueLanguageList.map((item) => {
-      return {
-        language: item,
-        count: allLanguageList.filter((language) => language === item).length,
-      };
-    });
-  };
-
-  const convertCountToPercentage = (count) => {
-    if (count > 10) {
-      return 100;
-    }
-    return count * 10;
-  };
+  const [sortedLanguageList, fetchRequestState, convertCountToPercentage] =
+    useSkills();
 
   return (
     <div id="skills">
@@ -57,12 +14,12 @@ export const Skills = () => {
           <h2>Skills</h2>
         </div>
         <div className="skills-container">
-          {state.requestState === requestStates.loading && (
+          {fetchRequestState === requestStates.loading && (
             <p className="description">取得中...</p>
           )}
-          {state.requestState === requestStates.success &&
-            state.languageList.map((item, index) => (
-              <div key={index}>
+          {fetchRequestState === requestStates.success &&
+            sortedLanguageList().map((item, index) => (
+              <div className="skill-item" key={index}>
                 <p className="description">
                   <strong>{item.language}</strong>
                 </p>
@@ -72,7 +29,7 @@ export const Skills = () => {
                 />
               </div>
             ))}
-          {state.requestState === requestStates.error && (
+          {fetchRequestState === requestStates.error && (
             <p className="description">エラーが発生しました</p>
           )}
         </div>
